change correct deleting for all entities and change response text for some api

+++ src/core/modules/categories/categories.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Param, Post, Delete, Put, Body } from '@nestjs/common';
+import { Controller, Get, Param, Post, Delete, Put, Body, BadRequestException } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 import { CategoriesService } from './categories.service';
 import { CreateCategoryDto } from './dto/create-category.dto';
@@ -7,33 +7,37 @@ import { Category } from 'src/database/models/category.model';
 @ApiTags('Categories')
 @Controller('categories')
 export class CategoriesController {
-  
-    constructor(private readonly categoriesService: CategoriesService) {}
+
+    constructor(private readonly categoriesService: CategoriesService) { }
 
     @Get('/allListCategories')
-    async getAll(){
-        return this.categoriesService.getAllCategory();
-      }
+    async getAll() {
+        return await this.categoriesService.getAllCategory();
+    }
 
- 
-     @Get(':id') 
-     getOne(@Param('id') id){
-          return this.categoriesService.getOneByIdCategory(id);
-      }
+
+    @Get(':id')
+    getOne(@Param('id') id): Promise<Category> {
+        return this.categoriesService.getOneByIdCategory(id);
+    }
 
     @Post('/create')
     async create(@Body() body: CreateCategoryDto): Promise<Category> {
         return this.categoriesService.createCategory(body);
-      }
-     
-      
+    }
+
+
     @Delete(':id')
-    remove(@Param('id') id): any{
-        return this.categoriesService.removeCategory(id);
+    async remove(@Param('id') id): Promise<string> {
+        const category = await this.categoriesService.getOneByIdCategory(id);
+        if (!category) throw new BadRequestException("Id invalid");
+
+        const resultDelete = await this.categoriesService.removeCategory(id);
+        return resultDelete == 1 ? "deletion successful " : "failed to delete"
     }
 
     @Put(':id')
-    edit(@Body() updateDto: UpdateCategoryDto, @Param('id') id: any) {
+    edit(@Body() updateDto: UpdateCategoryDto, @Param('id') id: number) {
         return this.categoriesService.editCategory(id, updateDto);
     }
