+++ src/core/modules/orders/orders.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Param, Delete, Post, Body, Put } from '@nestjs/common';
+import { Controller, Get, Param, Delete, Post, Body, Put, BadRequestException } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 import { OrdersService } from './orders.service';
 import { UpdateOrderDto } from './dto/update-order.dto';
@@ -10,33 +10,37 @@ import { Order } from 'src/database/models/order.model';
 export class OrdersController {
 
 
-    constructor(private readonly ordersService: OrdersService) {}
+    constructor(private readonly ordersService: OrdersService) { }
 
     @Get('/allListOrders')
-    async getAll(): Promise<Order[]>{
+    async getAll(): Promise<Order[]> {
         return this.ordersService.getAllOrders();
-      }
+    }
 
- 
-     @Get(':id') 
-     getOne(@Param('id') id:number): Promise<Order>{
-          return this.ordersService.getOneByIdOrder(id);
-      }
+
+    @Get(':id')
+    getOne(@Param('id') id: number): Promise<Order> {
+        return this.ordersService.getOneByIdOrder(id);
+    }
 
     @Post('')
     async create(@Body() body: CreateOrderDto): Promise<Order> {
         return this.ordersService.createOrder(body);
-      }
-     
-      
+    }
+
+
     @Delete(':id')
-    remove(@Param('id') id): string{
-        this.ordersService.removeOrder(id);
-    return `Delete order by ID: ${id}`
+    async remove(@Param('id') id:number): Promise<string> {
+        const order = await this.ordersService.getOneByIdOrder(id);
+        if (!order) throw new BadRequestException("Id invalid");
+
+        const resultDelete = await this.ordersService.removeOrder(id);
+        return resultDelete == 1 ? "deletion successful " : "failed to delete"
+
     }
 
     @Put(':id')
-    edit(@Body() updateDto: UpdateOrderDto, @Param('id') id: number){
+    edit(@Body() updateDto: UpdateOrderDto, @Param('id') id: number) {
         return this.ordersService.editOrder(id, updateDto);
     }
